chore: move LSM to dto
