+++ src/main/java/edu/kit/hci/soli/dto/LoginStateModel.java
@@ -1,5 +1,6 @@
-package edu.kit.hci.soli.domain;
+package edu.kit.hci.soli.dto;
 
+import edu.kit.hci.soli.domain.User;
 import org.jetbrains.annotations.Nullable;
 import org.springframework.security.web.csrf.CsrfToken;
