+++ src/main/jte/bookings.jte
@@ -1,4 +1,4 @@
-@import edu.kit.hci.soli.domain.LoginStateModel
+@import edu.kit.hci.soli.dto.LoginStateModel
 @import edu.kit.hci.soli.template.JteContext
 @import edu.kit.hci.soli.template.PageSpec
