+++ src/main/jte/error_known.jte
@@ -1,5 +1,5 @@
-@import edu.kit.hci.soli.domain.KnownError
-@import edu.kit.hci.soli.domain.LoginStateModel
+@import edu.kit.hci.soli.dto.KnownError
+@import edu.kit.hci.soli.dto.LoginStateModel
 @import edu.kit.hci.soli.template.JteContext
 
 @param JteContext context
