+++ src/main/jte/index.jte
@@ -1,4 +1,4 @@
-@import edu.kit.hci.soli.domain.LoginStateModel
+@import edu.kit.hci.soli.dto.LoginStateModel
 @import edu.kit.hci.soli.domain.Room
 @import edu.kit.hci.soli.template.JteContext
 @import edu.kit.hci.soli.template.PageSpec
